requirements.txt
openai
langchain
langchain-openai
chromadb
python-dotenv
